+++ .gitignore
@@ -1,4 +1,4 @@
-src/zonemap
+src/zns.fibmap
 
 Makefile
 configure
