+++ include/zmap.h
@@ -4,6 +4,7 @@
 #include <fcntl.h>
 #include <inttypes.h>
 #include <libgen.h>
+#include <linux/fiemap.h>
 #include <linux/blkzoned.h>
 #include <linux/fs.h>
 #include <stdio.h>
@@ -44,6 +45,27 @@ struct control {
 
 extern struct control ctrl;
 
+struct extent {
+    uint32_t zone;        /* zone index (starting with 1) of the extent */
+    uint32_t flags;       /* Flags given by ioctl() FIEMAP call */
+    uint32_t ext_nr;      /* Extent number as returned in the order by ioctl */
+    uint64_t logical_blk; /* LBA starting address of the extent */
+    uint64_t phy_blk;     /* PBA starting address of the extent */
+    uint64_t zone_lbas;   /* LBAS of the zone the extent is in */
+    uint64_t len;         /* Length of the extent in 512B sectors */
+    uint64_t zone_size;   /* Size of the zone the extent is in */
+    uint64_t zone_wp;     /* Write Pointer of this current zone */
+    uint64_t zone_lbae;   /* LBA that can be written up to (LBAS + ZONE CAP) */
+};
+
+struct extent_map {
+    uint32_t ext_ctr;  /* Number of extents in struct extent[] */
+    uint32_t zone_ctr; /* Number of zones in which extents are */
+    uint64_t
+        cum_extent_size;    /* Cumulative size of all extents in 512B sectors */
+    struct extent extent[]; /* Array of struct extent for each extent */
+};
+
 extern uint8_t is_zoned(char *);
 extern void init_dev(struct stat *);
 extern uint8_t init_znsdev(struct bdev *);
@@ -52,6 +74,10 @@ extern uint64_t get_zone_size(char *);
 extern void init_ctrl();
 extern uint32_t get_zone_number(uint64_t);
 extern void cleanup_ctrl();
+extern void print_zone_info(uint32_t zone);
+extern struct extent_map *get_extents();
+extern void sort_extents(struct extent_map *extent_map);
+extern void show_extent_flags(uint32_t flags);
 
 #define ERR_MSG(fmt, ...)                                                      \
     do {                                                                       \
@@ -67,7 +93,7 @@ extern void cleanup_ctrl();
 
 #define WARN(fmt, ...)                                                          \
     do {                                                                       \
-        printf("\033[0;33mWarning\033[0m: [%s:%4d] " fmt, \
+        printf("\033[0;33mWarning\033[0m: " fmt, \
                 __func__, __LINE__, ##__VA_ARGS__);            \
     } while (0)
