+++ src/fibmap.h
@@ -1,0 +1,8 @@
+#ifndef _FIBMAP_H_
+#define _FIBMAP_H_
+
+#include "zonemap.h"
+#include <inttypes.h>
+#include <stdint.h>
+
+#endif
